+++ .github/CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.1.3] - 2021-03-08
+### Added
+* `URL` argument type
+
 ## [0.1.2] - 2021-03-08
 ### Added
-* `*messageContext.SendMessageString` and `*messageContext.SendMessageEmbed` 
+* `*messageContext.SendMessageString` and `*messageContext.SendMessageEmbed`
 
 ## [0.1.1] - 2021-03-08
 ### Added
@@ -20,7 +24,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 * Everything (initial release)
 
-[Unreleased]: https://github.com/codemicro/dgo-toolbox/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/codemicro/dgo-toolbox/compare/v0.1.3...HEAD
+[0.1.3]: https://github.com/codemicro/dgo-toolbox/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/codemicro/dgo-toolbox/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/codemicro/dgo-toolbox/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/codemicro/dgo-toolbox/releases/tag/v0.1.0